fixed login bug,search and working on reminder

+++ logout.php
@@ -1,21 +1,32 @@
 <?php
 include 'inc/config/database.php';
 session_start();
-$userId=$_SESSION['id'];
-if (isset($_COOKIE['remember_me'])) {
-$cookie_parts = explode(':', $_COOKIE['remember_me']);
-$token = end($cookie_parts);
-    // Delete the user's session ID or token from the database
-    $query = "DELETE FROM remember_me_tokens WHERE user_id= ?";
-    $stmt = $pdo->prepare($query);
-    $stmt->execute([$userId]);
-    // Log the user out and delete the "remember me" cookie
-    //deleting the cookie
-    setcookie('remember_me', '', time() - 86400);
 
-    header("Location: login.php");
-}else{
-    header("Location: login.php");
+if (!empty($_SESSION['id'])) {
+    $userId = $_SESSION['id'];
+
+    if (isset($_COOKIE['remember_me'])) {
+        $cookie_parts = explode(':', $_COOKIE['remember_me']);
+        $token = end($cookie_parts);
+
+        // Validate and sanitize token
+        $token = filter_var($token, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
+
+        // Delete the user's token from the database
+        $query = "DELETE FROM remember_me_tokens WHERE user_id= ?";
+        $stmt = $pdo->prepare($query);
+        $stmt->execute([$userId]);
+
+        // If a token was found and deleted, delete the cookie
+        if ($stmt->rowCount() > 0) {
+            setcookie('remember_me', '', time() - 86400);
+        }
+    }
+
+    // Clear the session ID and log the user out
+    $_SESSION = array();
+    session_destroy();
 }
 
+header("Location: login.php");
 ?>
